Make it a two-parter

+++ install/fonts.sh
@@ -8,12 +8,16 @@ if ! fc-list | grep -qi "CaskaydiaMono Nerd Font"; then
   unzip CascadiaMono.zip -d CascadiaFont
   cp CascadiaFont/*.ttf ~/.local/share/fonts
   rm -rf CascadiaMono.zip CascadiaFont
+  fc-cache
+  cd -
+fi
 
+if ! fc-list | grep -qi "iA Writer Mono S"; then
+  cd /tmp
   wget -O iafonts.zip https://github.com/iaolo/iA-Fonts/archive/refs/heads/master.zip
   unzip iafonts.zip -d iaFonts
   cp iaFonts/iA-Fonts-master/iA\ Writer\ Mono/Static/iAWriterMonoS-*.ttf ~/.local/share/fonts
   rm -rf iafonts.zip iaFonts
-
   fc-cache
   cd -
 fi
